feature: esm, cjs, and types (#18)
* feature: esm, cjs, and types

#14 from the node-amqp-connnection-manager package they created a way to have both common and esm types being generated from TS code. Adapted their code to work with this. // moves back "types" into index.ts file so that it's generated correctly in dist on build

* feature: general cleanup

#14

* docs: fix capital letter
