+++ .github/workflows/ci.yaml
@@ -1,5 +1,11 @@
 name: GitHub CI
-on: [push, pull_request]
+on:
+  push:
+    branches: [master]
+    tags:
+      - '*'
+  pull_request:
+
 jobs:
   test:
     runs-on: ubuntu-latest
