+++ README.md
@@ -1,7 +1,9 @@
-# Fastify RabbitMq
+# Fastify RabbitMQ
 
 A Fastify RabbitMQ Plugin Developed in Pure TypeScript.
 It uses the [node-amqp-connection-manager](https://github.com/jwalton/node-amqp-connection-manager) plugin as a wrapper.
+
+The build exports this to vaild ESM and CJS for ease of cross-compatibility. 
 
 This comes right from the README on ```node-amqp-connection-manager```:
