+++ tsconfig.cjs.json
@@ -1,0 +1,9 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "declaration": false,
+        "sourceMap": false,
+        "module": "commonjs",
+        "outDir": "./dist/cjs"
+    }
+}
