+++ tsconfig.json
@@ -1,28 +1,30 @@
 {
-  "compileOnSave": true,
   "compilerOptions": {
-    "outDir": "./dist",
-    "allowJs": true,
-    "target": "es2015",
-    "module": "es2015",
-    "moduleResolution": "node",
+    "target": "es2018",
+    "module": "es6",
+    "lib": ["es2018"],
+    "allowJs": false,
     "declaration": true,
-    "allowUnreachableCode": false,
-    "noImplicitReturns": true,
+    "sourceMap": true,
+    "outDir": "./dist/esm",
+    "stripInternal": true,
+    "strict": true,
     "noImplicitAny": true,
-    "esModuleInterop": true,
+    "noImplicitThis": true,
+    "alwaysStrict": true,
+    "forceConsistentCasingInFileNames": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noImplicitReturns": true,
+    "noFallthroughCasesInSwitch": true,
+    "moduleResolution": "node",
     "strictNullChecks": true,
-    "types": ["jest", "node"]
+    "typeRoots": [
+      "./node_modules/@types",
+      "./@types"
+    ],
+    "esModuleInterop": true
   },
-  "include": [
-    "./**/*"
-  ],
-  "exclude": [
-    "node_modules",
-    "jest.config.ts",
-    "./__tests__/**/*",
-    "./coverage/**/*",
-    "./types/**/*",
-    "./dist/**/*"
-  ]
+  "include": ["./src/**/*"],
+  "compileOnSave": true
 }
